move reused model manipulation into architects.py
